Added Past Orders page.

+++ db/migrate/20211114165649_create_card_orders.rb
@@ -4,7 +4,7 @@ class CreateCardOrders < ActiveRecord::Migration[6.1]
       t.references :order, null: false, foreign_key: true
       t.references :card, null: false, foreign_key: true
       t.integer :quantity
-      t.integer :price
+      t.decimal :price
 
       t.timestamps
     end
